Refactor admin dashboard to use server-side data fetching with client component
